Move 'blend_palette'.c to util.c
