--- include/blend_palette.h
@@ -1,6 +1,0 @@
-#ifndef GUARD_BLEND_PALETTE_H
-#define GUARD_BLEND_PALETTE_H
-
-void BlendPalette(u16 palOffset, u16 numEntries, u8 coeff, u16 blendColor);
-
-#endif // GUARD_BLEND_PALETTE_H
