+++ include/util.h
@@ -13,5 +13,6 @@ int CountTrailingZeroBits(u32 value);
 u16 CalcCRC16(u8 *data, s32 length);
 u16 CalcCRC16WithTable(u8 *data, u32 length);
 u32 CalcByteArraySum(const u8* data, u32 length);
+void BlendPalette(u16 palOffset, u16 numEntries, u8 coeff, u16 blendColor);
 
 #endif // GUARD_UTIL_H
