+++ src/battle_anim_80A5C6C.c
@@ -3,7 +3,6 @@
 #include "constants/species.h"
 #include "battle.h"
 #include "battle_anim.h"
-#include "blend_palette.h"
 #include "contest.h"
 #include "data2.h"
 #include "decompress.h"
