+++ src/util.c
@@ -1,6 +1,7 @@
 #include "global.h"
 #include "util.h"
 #include "sprite.h"
+#include "palette.h"
 
 const u32 gBitTable[] =
 {
@@ -257,4 +258,21 @@ u32 CalcByteArraySum(const u8* data, u32 length)
     for (sum = 0, i = 0; i < length; i++)
         sum += data[i];
     return sum;
+}
+
+void BlendPalette(u16 palOffset, u16 numEntries, u8 coeff, u16 blendColor)
+{
+    u16 i;
+    for (i = 0; i < numEntries; i++)
+    {
+        u16 index = i + palOffset;
+        struct PlttData *data1 = (struct PlttData *)&gPlttBufferUnfaded[index];
+        s8 r = data1->r;
+        s8 g = data1->g;
+        s8 b = data1->b;
+        struct PlttData *data2 = (struct PlttData *)&blendColor;
+        gPlttBufferFaded[index] = ((r + (((data2->r - r) * coeff) >> 4)) << 0)
+                                | ((g + (((data2->g - g) * coeff) >> 4)) << 5)
+                                | ((b + (((data2->b - b) * coeff) >> 4)) << 10);
+    }
 }
